feat(go): Edit request-review to use lib
